Fixes #290
This commit removes an optimization that stops suggestion rendering
when further keys are queued for processing. This optimization can lead
to wrong results because there is no guarantee that the last queued key
will invoke the `modify` widget (it could be a vi movement for example).
In that case, no suggestion is rendered.